requirements.txt
kafka-python
pymongo
fastapi
uvicorn
strawberry-graphql
faker
pandas
prometheus_client